+++ lib/src/models/article_basket.dart
@@ -5,7 +5,7 @@ import 'package:models_weebi/base.dart';
 
 import 'package:models_weebi/src/models/proxy_article_worth.dart';
 import 'package:models_weebi/utils.dart';
-import 'package:models_weebi/weebi_models.dart' show ProxyArticle, ArticleLines;
+import 'package:models_weebi/weebi_models.dart' show ProxyArticle, ArticleLine;
 import 'package:collection/collection.dart';
 
 class ArticleBasket extends ArticleAbstract {
@@ -40,7 +40,7 @@ class ArticleBasket extends ArticleAbstract {
         );
 
   Iterable<ProxyArticleWorth> getProxiesListWithPriceAndCost(
-      Iterable<ArticleLines> lines) {
+      Iterable<ArticleLine> lines) {
     final proxiesWorth = <ProxyArticleWorth>[];
     for (final p in proxies) {
       final temp = p.getProxyArticleWorth(lines);
@@ -51,7 +51,7 @@ class ArticleBasket extends ArticleAbstract {
 
   static Iterable<ProxyArticleWorth>
       getProxiesListWithPriceAndCostArticleNotCreatedYetOnly(
-          Iterable<ArticleLines> lines, Iterable<ProxyArticle> proxiesRaw) {
+          Iterable<ArticleLine> lines, Iterable<ProxyArticle> proxiesRaw) {
     final proxiesWorth = <ProxyArticleWorth>[];
     for (final p in proxiesRaw) {
       final temp = p.getProxyArticleWorth(lines);