+++ test/weebi_models/article_line_test.dart
@@ -1,13 +1,13 @@
-import 'package:models_weebi/src/models/articles_lines.dart';
+import 'package:models_weebi/src/models/article_line.dart';
 import 'package:test/test.dart';
 
 void main() {
   test('line article weebi json', () {
-    final json = ArticleLines.dummy.toJson();
-    final dummy = ArticleLines.fromJson(json);
+    final json = ArticleLine.dummy.toJson();
+    final dummy = ArticleLine.fromJson(json);
     // expect(dummy == ProductWeebi.dummy, isTrue);
-    expect(dummy.id == ArticleLines.dummy.id, isTrue);
-    expect(dummy.status == ArticleLines.dummy.status, isTrue);
-    expect(dummy.creationDate == ArticleLines.dummy.creationDate, isTrue);
+    expect(dummy.id == ArticleLine.dummy.id, isTrue);
+    expect(dummy.status == ArticleLine.dummy.status, isTrue);
+    expect(dummy.creationDate == ArticleLine.dummy.creationDate, isTrue);
   });
 }